backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
langgraph>=0.1.0
langchain-core>=0.2.0
motor>=3.4.0          # async MongoDB driver
httpx>=0.27.0
google-genai
python-dotenv>=1.0.0
pydantic>=2.7.0
mangum>=0.17.0        # AWS Lambda adapter for FastAPI
